+++ catsmat/libIMUSANT/Utilities/inc/TRational.h
@@ -38,23 +38,23 @@ namespace IMUSANT
         
     private:
         
-        long int fNumerator;
-        long int fDenominator;
+        long fNumerator;
+        long fDenominator;
         
         /// Used by rationalise()
-        long int gcd(long int a, long int b);
+        long int gcd(long a, long b);
         
     public:
         
-        TRational(long int num = 0, long int denom = 1);
+        TRational(long num = 0, long denom = 1);
         TRational(const TRational& d);
         TRational(const string &str);
         
-        long int getNumerator()	const		{ return fNumerator; }
-        long int getDenominator() const		{ return fDenominator; }
-        void setNumerator(long int d)		{ fNumerator = d; }
-        void setDenominator(long int d) 	{ fDenominator = d; }
-        void set(long int n, long int d)  { fNumerator = n; fDenominator = d; }
+        long getNumerator()	const		{ return fNumerator; }
+        long getDenominator() const		{ return fDenominator; }
+        void setNumerator(long d)		{ fNumerator = d; }
+        void setDenominator(long d) 	{ fDenominator = d; }
+        void set(long int n, long d)  { fNumerator = n; fDenominator = d; }
         
         TRational operator +(const TRational &dur) const;
         TRational operator -(const TRational &dur) const;
@@ -63,8 +63,8 @@ namespace IMUSANT
         TRational operator /(const TRational &dur) const;
         // (i.e. dur * 3/2 or dur * 7/4)
         
-        TRational operator *(int num) const;
-        TRational operator /(int num) const;
+        TRational operator *(long num) const;
+        TRational operator /(long num) const;
         
         TRational& operator +=(const TRational &dur);
         TRational& operator -=(const TRational &dur);
@@ -73,8 +73,8 @@ namespace IMUSANT
         TRational& operator /=(const TRational &dur);
         // (i.e. dur * 3/2 or dur * 7/4)
         
-        TRational& operator *=(long int num) { fNumerator *= num; return *this; }
-        TRational& operator /=(long int num) { fDenominator *= num; return *this; }
+        TRational& operator *=(long num) { fNumerator *= num; return *this; }
+        TRational& operator /=(long num) { fDenominator *= num; return *this; }
         
         TRational& operator =(const TRational& dur);
